update: refactor multiple cifar to class based view
